#신입-베이스캠프-8기-루키여우TF/1 자료형 확인 부탁
누구야!!!

+++ main.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
  
-double add(double, int);
-double sub(double, int);
-double div(double, int);
-double multi(double, int);
+double add(double, double);
+double sub(double, double);
+double div(double, double);
+double multi(double, double);
  
 int main(int argc, char* argv[])
 {
